EduProgramProfileForm: It's sufficent to have single int field for PK - add unique constraint for FK fields instead, drop duplicating index

+++ R7.University.Controls/ViewModels/EduProgramProfileFormEditModel.cs
@@ -61,7 +61,7 @@ namespace R7.University.Controls.ViewModels
 
         #region IEduProgramProfileFormWritable implementation
 
-        public long EduProgramProfileFormID { get; set; }
+        public int EduProgramProfileFormID { get; set; }
 
         public int EduProgramProfileID { get; set; }
